+++ app/src/main/java/com/theuhooi/uhooipicbook/modules/monsterlist/dummy/DummyContent.kt
@@ -1,4 +1,4 @@
-package com.theuhooi.uhooipicbook.monsterlist.dummy
+package com.theuhooi.uhooipicbook.modules.monsterlist.dummy
 
 import java.util.*
 
@@ -25,7 +25,11 @@ object DummyContent {
     init {
         // Add some sample items.
         for (i in 1..COUNT) {
-            addItem(createDummyItem(i))
+            addItem(
+                createDummyItem(
+                    i
+                )
+            )
         }
     }
 
@@ -35,7 +39,13 @@ object DummyContent {
     }
 
     private fun createDummyItem(position: Int): DummyItem {
-        return DummyItem(position.toString(), "Item $position", makeDetails(position))
+        return DummyItem(
+            position.toString(),
+            "Item $position",
+            makeDetails(
+                position
+            )
+        )
     }
 
     private fun makeDetails(position: Int): String {
